requirements.txt
python-dotenv
langchain
openai
pyautogui
pygetwindow
pyperclip
keyboard
